requirements.txt
amqp>=2.5.2
billiard>=3.6.3.0
Click==7.0
Flask==1.1.1
kombu>=4.6.8
more-itertools==7.2.0
pytz==2019.3
redis==3.3.8
vine==1.3.0
Werkzeug==0.16.0
zipp==0.6.0
Flask-Script==2.0.6
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
configparser==4.0.2
extend-me==1.1.4
idna==2.8
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
MarkupSafe==1.1.1
Paste==3.4.0
py-bcrypt==0.4
pyrsistent==0.15.7
requests==2.22.0
simplejson==3.17.0
six==1.14.0
urllib3==1.25.8
waitress==1.4.3
PyJWT==1.7.1
pika
google-cloud-secret-manager==0.2.0
Flask-WTF==0.14.3
